scheduleModel調整
