+++ NiccaScheduler/Model/TaskModel.swift
@@ -35,21 +35,3 @@ class TaskSchedulePageCountModel: Object {
     @objc dynamic var createdAt = Date()
     @objc dynamic var page1DayCount: Int = 0 // 現在の１日に予定されているページ数
 }
-
-// タスクごと日毎スケジュール
-class TaskScheduleModel: Object {
-    @objc dynamic var taskScheduleId: String = NSUUID().uuidString // タスクスケジュールID primarykey
-    @objc dynamic var taskId: String? = nil // タスクID relationkey
-    @objc dynamic var scheduleStartPageNumber: Int = 0 // 今日スタート予定のページ数
-    @objc dynamic var scheduleEndPageNumber: Int = 0 // 今日終了予定のページ数
-    @objc dynamic var endedPageNumber: Int = 0 // 今日実際に終了したページ数(やってない場合はstartPage - 1)
-    @objc dynamic var createdAt = Date()
-    @objc dynamic var updatedAt = Date()
-    @objc dynamic var executionDate: Date? = nil // 実行する日時 YYYY-MM-DD
-    
-    let task = LinkingObjects(fromType: TaskModel.self, property: "taskSchedules")
-    
-    override static func primaryKey() -> String? {
-        return "taskScheduleId"
-    }
-}
