+++ NiccaScheduler/Model/TaskScheduleModel.swift
@@ -1,18 +1,29 @@
 //
-//  CurrentTaskSchedule.swift
+//  DoneSchedule.swift
 //  NiccaScheduler
 //
-//  Created by 北洞亜也加 on 2021/08/18.
+//  Created by 北洞亜也加 on 2021/09/01.
 //
 
 import Foundation
 import RealmSwift
 
-class CurrentTaskScheduleModel: Object {
-    dynamic var currentTaskScheduleList: Results<TaskScheduleModel>!
-    @objc dynamic var currentDay: Date!
-    @objc dynamic var selectedDate: String!
-    dynamic var TaskListResluts: Results<TaskModel>!
+// タスクごと日毎スケジュール
+class TaskScheduleModel: Object {
+    @objc dynamic var taskScheduleId: String = NSUUID().uuidString // タスクスケジュールID primarykey
+    @objc dynamic var taskId: String? = nil // タスクID relationkey
+    @objc dynamic var scheduleStartPageNumber: Int = 0 // 今日スタート予定のページ数
+    @objc dynamic var scheduleEndPageNumber: Int = 0 // 今日終了予定のページ数
+    @objc dynamic var endedPageNumber: Int = 0 // 今日実際に終了したページ数(やってない場合はstartPage - 1)
+    @objc dynamic var createdAt = Date()
+    @objc dynamic var updatedAt = Date()
+    @objc dynamic var executionDate: Date? = nil // 実行する日時 YYYY-MM-DD
+    
+    let task = LinkingObjects(fromType: TaskModel.self, property: "taskSchedules")
+    
+    override static func primaryKey() -> String? {
+        return "taskScheduleId"
+    }
     
     override required init() {
         super.init()
@@ -20,31 +31,36 @@ class CurrentTaskScheduleModel: Object {
     }
 }
 
-extension CurrentTaskScheduleModel {
+extension TaskScheduleModel {
 
+    static func getDoneTaskSchedule(with date: Date) -> Results<TaskScheduleModel>! {
+        let RealmInstance = try! Realm()
+        return RealmInstance.objects(TaskScheduleModel.self).filter("endedPageNumber > 0")
+    }
+    
     static func createTaskSchedule(with date: Date) -> Results<TaskScheduleModel>! {
         let model = self.init()
         
         let calendar = Calendar.current
         let startTime = calendar.startOfDay(for: date)
 
-        model.currentDay = startTime
-        model.currentTaskScheduleList = nil
-        model.selectedDate = DateUtils.stringFromDate(date: model.currentDay, format: "YYYY-MM-dd")
-        model.TaskListResluts = nil
+        let currentDay = startTime
+        let currentTaskScheduleList: Results<TaskScheduleModel>!
+        let selectedDate = DateUtils.stringFromDate(date: currentDay, format: "YYYY-MM-dd")
+        let TaskListResluts: Results<TaskModel>!
         
         let RealmInstance = try! Realm()
-        model.TaskListResluts = RealmInstance.objects(TaskModel.self) // TODO: 期間中のタスクに絞る
-        if (model.TaskListResluts.count > 0) {
-            for task in model.TaskListResluts {
-                let taskScheduleList = RealmInstance.objects(TaskScheduleModel.self).filter("executionDate == %@", model.currentDay as Any).filter("taskId == %@", task.taskId)
+        TaskListResluts = RealmInstance.objects(TaskModel.self) // TODO: 期間中のタスクに絞る
+        if (TaskListResluts.count > 0) {
+            for task in TaskListResluts {
+                let taskScheduleList = RealmInstance.objects(TaskScheduleModel.self).filter("executionDate == %@", currentDay as Any).filter("taskId == %@", task.taskId)
                 // 一旦その日のスケジュールがない場合に作成
                 // 終わったページを更新したタイミングで作成しなおす必要がある
                 if (taskScheduleList.count == 0) {
                     // その日の分のタスクスケジュールを作成する
                     // TODO: 一旦 page1DayCountから割って算出するけど、実際の前日終了ページ数(endedPageNumber)から計算しなおす必要がある
                     // 何日間で行うか
-                    let dayInterval = Int((Calendar.current.dateComponents([.day], from: task.scheduleStartAt!, to: model.currentDay)).day!)
+                    let dayInterval = Int((Calendar.current.dateComponents([.day], from: task.scheduleStartAt!, to: currentDay)).day!)
                     print("dayInterval")
                     print(dayInterval)
                     
@@ -76,7 +92,7 @@ extension CurrentTaskScheduleModel {
                     instanceTaskScheduleModel.taskId = task.taskId
                     instanceTaskScheduleModel.scheduleStartPageNumber = scheduleStartPageNumber
                     instanceTaskScheduleModel.scheduleEndPageNumber = scheduleEndPageNumber
-                    instanceTaskScheduleModel.executionDate = model.currentDay
+                    instanceTaskScheduleModel.executionDate = currentDay
                     
                     try! RealmInstance.write {
                         task.taskSchedules.append(instanceTaskScheduleModel)
@@ -86,7 +102,8 @@ extension CurrentTaskScheduleModel {
             }
         }
         // スケジュールを取得し直す
-        model.currentTaskScheduleList = RealmInstance.objects(TaskScheduleModel.self).filter("executionDate == %@", model.currentDay as Any)
-        return model.currentTaskScheduleList
+        return  RealmInstance.objects(TaskScheduleModel.self).filter("executionDate == %@", currentDay as Any)
     }
 }
+
+
